refactor: expose graph builders for eval
