+++ eval/variants.yaml
@@ -6,7 +6,7 @@ variants:
       model: "llama3.1:8b-instruct"
       temperature: 0.0
     prompt: "eval/prompts/default.j2"
-    graph: "assist.eval_graphs:reflexion_graph_v1"
+    graph: "assist.reflexion_agent:reflexion_graph_v1"
 
   - name: planner_default
     model: openai_compatible
@@ -15,7 +15,7 @@ variants:
       model: "llama3.1:8b-instruct"
       temperature: 0.0
     prompt: "eval/prompts/default.j2"
-    graph: "assist.eval_graphs:planner_graph_v1"
+    graph: "assist.reflexion_agent:planner_graph_v1"
 
   - name: plan_checker_default
     model: openai_compatible
@@ -24,7 +24,7 @@ variants:
       model: "llama3.1:8b-instruct"
       temperature: 0.0
     prompt: "eval/prompts/default.j2"
-    graph: "assist.eval_graphs:plan_checker_graph_v1"
+    graph: "assist.reflexion_agent:plan_checker_graph_v1"
 
   - name: step_executor_default
     model: openai_compatible
@@ -33,7 +33,7 @@ variants:
       model: "llama3.1:8b-instruct"
       temperature: 0.0
     prompt: "eval/prompts/default.j2"
-    graph: "assist.eval_graphs:step_executor_graph_v1"
+    graph: "assist.reflexion_agent:step_executor_graph_v1"
 
   - name: summarizer_default
     model: openai_compatible
@@ -42,4 +42,4 @@ variants:
       model: "llama3.1:8b-instruct"
       temperature: 0.0
     prompt: "eval/prompts/default.j2"
-    graph: "assist.eval_graphs:summarizer_graph_v1"
+    graph: "assist.reflexion_agent:summarizer_graph_v1"
